nit: create file linera-storage/src/db_store.rs
